Finish TODOs 🎉 (#37)
* Encode external funs

* Avoid re-boxing

* Fix copy/pasta

* Delete TODOs

After reading a bunch of docs, it seems like you never want to put PIDs,
Ports, References, or internal `fun`s inside a BEAM file. They are part
of the External Term Format because they get used for inter-process
communication. I should probably verify this with someone that knows
more, but it seems safe to exclude them until they are asked for.

Resources:
 - https://gist.github.com/physacco/5802293
 - https://github.com/lindenbaum/hinterface/blob/master/src/Foreign/Erlang/Term.hs#L585
 - https://github.com/erlang/otp/blob/83e20c62057ebc1d8064bf57b01be560cd244e1d/lib/erl_interface/test/erl_eterm_SUITE.erl#L1069

* Generalize export metadata

* Refactor Module data to function of env

* Add \`insertModuleInfo\`

* Make module_info insertion safe

* Move functionCount